más correcciones refactor arquitectura
