+++ tpi/src/main/java/com/unlam/tpi/delivery/controlador/HistoricoControladorImpl.java
@@ -1,7 +1,7 @@
-package com.unlam.tpi.controlador;
+package com.unlam.tpi.delivery.controlador;
 
 import com.unlam.tpi.modelo.rest.FechaRequestHistorico;
-import com.unlam.tpi.interfaces.HistoricoServicio;
+import com.unlam.tpi.core.interfaces.HistoricoServicio;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
