+++ tpi/src/main/java/com/unlam/tpi/infraestructura/repositorio/HistoricoRepositorio.java
@@ -1,4 +1,4 @@
-package com.unlam.tpi.repositorio;
+package com.unlam.tpi.infraestructura.repositorio;
 
 import java.util.List;
